requirements.txt
# Minimal requirements for Vercel deployment
# Using ONNX Runtime (lightweight inference engine)

numpy==1.24.3
Pillow==10.0.0
flask==3.0.0
werkzeug==3.0.0
onnxruntime==1.16.3

# Note: You'll need to convert your .h5 model to .onnx format
# See DEPLOYMENT.md for conversion instructions
